+++ SleepHunter/Properties/AssemblyInfo.cs
@@ -55,4 +55,4 @@ using System.Windows;
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("4.4.1.0")]
 [assembly: AssemblyFileVersion("4.4.1.0")]
-[assembly: AssemblyInformationalVersion("23415")]
+[assembly: AssemblyInformationalVersion("23416")]