+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.4.1] - Unreleased
+## [4.4.1] - 2023-04-16
 
 ### Added
 
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Disable spell queue remove buttons on startup (when empty)
 - SleepHunter window title not changing sometimes
+- Disable skill/spell tabs when no client
+- Disable spell queue when no client
 
 ## [4.4.0] - 2023-04-15